requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
networkx>=3.0.0
imbalanced-learn>=0.11.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
loguru>=0.7.3